requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
discord==2.3.2
discord-py==2.4.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google-gemini/generative-ai-python@2786f33edbabc441b247e549025b3e9e5e849895
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
playwright==1.49.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.4
pydantic-core==2.27.2
pyee==12.0.0
pyparsing==3.2.1
python-weather==2.0.7
pytubefix==8.8.4
requests==2.32.3
rsa==4.9
soupsieve==2.6
tqdm==4.67.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wavelink==3.4.1
yarl==1.18.3